Move modules related to parse packets to internal
